move mapArguments to atrium-logic
- deprecate the one in domain-builders
- update usages
- update ReadmeSpec (and in consequence README.md)
